Update: 1. add const definition for alert methods. 2. use putToSendClass function for putting alerts to SendClass. 3. move methods Send*** to send.go.
