+++ cmd/alert-gateway/common/const.go
@@ -1,0 +1,8 @@
+package common
+
+const (
+	AlertMethodSms    = "SMS"
+	AlertMethodLanxin = "LANXIN"
+	AlertMethodCall   = "CALL"
+	AlertMethodHook   = "HOOK"
+)
